Begin Gradle modularization
